Put DisplayRecipe function in display.go. (#24)
